buildbot: adjust upload paths in packages.py
This commit introduces a symlink-setup: The artifacts get uploaded to a
separate directory, from where they get linked to the final destination
by symlinks. Linking the new artifacts happens only, if the complete build
was successful.

That way, we keep the old artifacts, in case something went wrong.

Should fix #28