buildbot: start the build containers with sudo
Now that we have workers running in containers (on podone & podtwo),
we need to be more deliberate in how the containers are constructed
and which capabilities and devices we give them.

I already made sure we don't need --privileged on the outer container,
but this sudo on the inner container is still needed
until Podman and OpenWrt integrate more tightly.